many changes

+++ utils/clearUploadDirectory.js
@@ -7,9 +7,18 @@ import fs from 'fs';
  * @param {object} req - Express request object
  */
 export const clearUploadDirectory = (req) => {
-  if (req.files && req.files.length > 0) {
-    req.files.forEach((file) => {
-      fs.unlinkSync(file.path);
-    });
+  try {
+    if (req.files && req.files.length > 0) {
+      req.files.forEach((file) => {
+        if (fs.existsSync(file.path)) {
+          fs.unlinkSync(file.path);
+        }
+      });
+    } else if (req.file) {
+      fs.unlinkSync(req.file.path);
+    }
+  } catch (error) {
+    // Handle the error
+    console.error('Error while clearing upload directory:', error);
   }
 };
